Remove deprecated and rework reverse
